Isolate connect button and move to service header
