+++ frontend/src/components/Connect.tsx
@@ -1,5 +1,5 @@
-import React, { useState, useEffect } from 'react'
-import { List, Button, Typography, Collapse, Divider } from '@mui/material'
+import React, { useEffect } from 'react'
+import { List, Button, Collapse, Divider } from '@mui/material'
 import { ApplicationState, Dispatch } from '../store'
 import { useSelector, useDispatch } from 'react-redux'
 import { useParams, useLocation } from 'react-router-dom'
@@ -7,10 +7,9 @@ import { canUseConnectLink } from '../models/applicationTypes'
 import { getDeviceModel } from '../selectors/devices'
 import { DeviceContext } from '../services/Context'
 import { windowOpen } from '../services/Browser'
-import { makeStyles } from '@mui/styles'
 import { LicensingNotice } from './LicensingNotice'
 import { ConnectionDetails } from './ConnectionDetails'
-import { ConnectionErrorMessage } from './ConnectionErrorMessage'
+import { ServiceConnectButton } from '../buttons/ServiceConnectButton'
 import { ConnectLinkSetting } from './ConnectLinkSetting'
 import { ServiceAttributes } from './ServiceAttributes'
 import { NetworksAccordion } from './NetworksAccordion'
@@ -20,20 +19,13 @@ import { AutoLaunchToggle } from './AutoLaunchToggle'
 import { ConnectionSurvey } from './ConnectionSurvey'
 import { AccessAccordion } from './AccessAccordion'
 import { ConnectionMenu } from './ConnectionMenu'
-import { ComboButton } from '../buttons/ComboButton'
-import { GuideBubble } from './GuideBubble'
-import { ErrorButton } from '../buttons/ErrorButton'
-import { DesktopUI } from './DesktopUI'
-import { DataCopy } from './DataCopy'
 import { PortalUI } from './PortalUI'
 import { Gutters } from './Gutters'
-import { spacing } from '../styling'
 import { Notice } from './Notice'
 import { Icon } from './Icon'
 import { Pre } from './Pre'
 
 export const Connect: React.FC = () => {
-  const css = useStyles()
   const location = useLocation<{
     autoConnect?: boolean
     autoLaunch?: boolean
@@ -42,18 +34,13 @@ export const Connect: React.FC = () => {
   }>()
   const { sessionID } = useParams<{ deviceID?: string; sessionID?: string }>()
   const { connection, device, service, instance } = React.useContext(DeviceContext)
-  const [showError, setShowError] = useState<boolean>(true)
-  const [connectThisDevice, setConnectThisDevice] = useState<boolean>(false)
   const dispatch = useDispatch<Dispatch>()
-  const { session, accordion, thisDevice, showConnectLink } = useSelector((state: ApplicationState) => ({
-    thisDevice: device?.thisDevice && instance?.owner.id === state.user.id,
+  const { session, accordion, showConnectLink } = useSelector((state: ApplicationState) => ({
     session: state.sessions.all.find(s => s.id === sessionID),
     fetching: getDeviceModel(state).fetching,
     accordion: state.ui.accordion,
     showConnectLink: canUseConnectLink(state, service?.typeID) && !(!connection.connectLink && connection.enabled),
   }))
-
-  const displayThisDevice = thisDevice && !connectThisDevice
 
   useEffect(() => {
     if (!location.state) return
@@ -66,21 +53,6 @@ export const Connect: React.FC = () => {
 
   return (
     <>
-      {displayThisDevice && (
-        <Notice gutterTop solid onClose={() => setConnectThisDevice(true)}>
-          <Typography variant="h3">You are on the same device as this service.</Typography>
-          <Typography variant="body2" gutterBottom>
-            So you can connect directly without using Remote.It.
-          </Typography>
-          <DataCopy
-            label="Local endpoint"
-            value={`${service.host || '127.0.0.1'}:${service.port}`}
-            showBackground
-            alwaysWhite
-            fullWidth
-          />
-        </Notice>
-      )}
       <ConnectionDetails
         connection={connection}
         service={service}
@@ -88,56 +60,7 @@ export const Connect: React.FC = () => {
         show={!!(connection.enabled && connection.host) || connection.connectLink}
       />
       {service.license === 'UNLICENSED' && <LicensingNotice instance={instance} />}
-      {!displayThisDevice && !connection.connectLink && (
-        <GuideBubble
-          guide="connectButton"
-          enterDelay={400}
-          placement="left"
-          startDate={new Date('2022-09-20')}
-          queueAfter={device ? 'availableServices' : 'addNetwork'}
-          instructions={
-            <>
-              <Typography variant="h3" gutterBottom>
-                <b>
-                  <PortalUI>Starting a connection</PortalUI>
-                  <DesktopUI>Connect on demand</DesktopUI>
-                </b>
-              </Typography>
-              <PortalUI>
-                <Typography variant="body2" gutterBottom>
-                  Create a connection to this service with the button to the right.
-                </Typography>
-              </PortalUI>
-              <DesktopUI>
-                <Typography variant="body2" gutterBottom>
-                  Start listening on to this endpoint for network requests. On request, automatically create the
-                  connection and disconnect when idle.
-                </Typography>
-              </DesktopUI>
-              {connection.autoLaunch && (
-                <Typography variant="body2" gutterBottom>
-                  <em>
-                    This connection will launch when connected because the "Auto Launch" configuration toggle is on.
-                  </em>
-                </Typography>
-              )}
-            </>
-          }
-        >
-          <Gutters size="md" className={css.gutters} bottom={null}>
-            <ErrorButton connection={connection} onClick={() => setShowError(!showError)} visible={showError} />
-            <ComboButton
-              size="large"
-              service={service}
-              connection={connection}
-              permissions={instance.permissions}
-              onClick={() => dispatch.ui.guide({ guide: 'aws', step: 6 })}
-              fullWidth
-            />
-          </Gutters>
-        </GuideBubble>
-      )}
-      <ConnectionErrorMessage connection={connection} visible={showError} />
+      {/* <ServiceConnectButton /> */}
       <ConnectionSurvey connection={connection} highlight={!!location.state?.autoFeedback} />
       <Gutters size="md" bottom={null}>
         <AccordionMenuItem
@@ -161,11 +84,8 @@ export const Connect: React.FC = () => {
             </Collapse>
             <PortalUI>
               <Notice gutterTop severity="info">
-                <strong>Want Persistent URLs for your system?</strong>
-                <em>
-                  Get Peer-to-peer on demand connections with persistent URLs. Remote system and network access.
-                  Improved launch commands and the Remote.It CLI.
-                </em>
+                <strong>Want Persistent URLs?</strong>
+                <em>On demand connections with persistent URLs, remote access, and improved launch commands.</em>
                 <Button
                   size="small"
                   color="primary"
@@ -205,13 +125,3 @@ export const Connect: React.FC = () => {
     </>
   )
 }
-
-const useStyles = makeStyles({
-  actions: {
-    marginRight: spacing.md,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  gutters: { display: 'flex', alignItems: 'flex-end', '& button': { height: 45 } },
-})
